+++ composeApp/src/wasmJsMain/kotlin/org/noiseplanet/noisecapture/services/storage/OPFSFileSystemService.kt
@@ -4,10 +4,13 @@ import kotlinx.browser.document
 import kotlinx.coroutines.await
 import org.noiseplanet.noisecapture.interop.ZipJs
 import org.noiseplanet.noisecapture.interop.storage.FileSystemWritableFileStream
+import org.noiseplanet.noisecapture.util.FeatureCollection
+import org.noiseplanet.noisecapture.util.GeoJson
 import org.noiseplanet.noisecapture.util.OPFSHelper
 import org.w3c.dom.HTMLAnchorElement
 import org.w3c.dom.url.URL
 import org.w3c.files.Blob
+import org.w3c.files.BlobPropertyBag
 import org.w3c.files.File
 
 /**
@@ -65,6 +68,16 @@ class OPFSFileSystemService : FileSystemService {
 
         // Download zip file
         downloadBlob(blob, "$archiveName.zip")
+    }
+
+    override suspend fun downloadGeoJson(geoJson: FeatureCollection, fileName: String) {
+        val contents = GeoJson.encodeToString(geoJson)
+        // Write data to blob
+        val blob = Blob(
+            blobParts = arrayOf<JsAny?>(contents.toJsString()).toJsArray(),
+            options = BlobPropertyBag(type = "application/geo+json")
+        )
+        downloadBlob(blob, fileName)
     }
 
     /**